requirements.txt
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
waitress==3.0.0
